+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
   [gbastien]
 - Improved `transmogrifier.filter_keys`
   [sgeulette]
+- Added `workflow.update_role_mappings_for` helper to update WF role mappings
+  for a given object.
+  [gbastien]
 
 0.69 (2023-05-31)
 -----------------